My Own Texture Placement Class
My texture placement takes the current uv, no matter what it is, and applies the TRS calculations for textures. This class is a better
design than Alex's in the world of MVC since the controller only needs to know how to pass floats and vector2's. Does not need to know
about Matrix implementation or concerning itself with TRS.